reformatted files js -> jsx, css -> scss, added relevant and removed obfuscatory commenting, straightened out the indentations.
