+++ client/components/user/nav-bar.jsx
@@ -1,38 +1,40 @@
 import React from 'react';
 /* 
 To Do: 
-    -nav bar gets location and can pass to restaurant and events
+    - turn nav bar into actual bar
+    - add logout button
+    - add home button?
+    - lots of styling on main app pages
 */
 const NavBar = (props) => {
     return (
         // nav bar, with logout button and location query
         <nav>
             <form
+            // on form submission, ingests the location from the location textbox on the page
                 onSubmit={(e) => {
                     e.preventDefault();
                     const location = e.target[0].value;
-                    // make fetch post request to backend
+                    // makes a post request to api/location with route parameter location converted to JSON
                     fetch(`/api/location/${location}`, {
                         method: 'POST',
                         headers: { 'Content-Type': 'application/json' },
                         body: JSON.stringify({ location }),
                     })
                         .then((response) => {
-                            
-                            // console.log('this should be initial response:', response);
+                            // returns the results of the backend call to locationController.getEvents
+                            // and locationController.getRestaurants in JSON format (then parsed)
                             return response.json();
                         })
                         .then((json) => {
-                            // save data and pass it to restaurant container. will be X amount of restaurants.
-                            // console.log('this should be the location response to json', json);
-                            // update location here
-                            // console.log('this should be response jsonified', json)
-                            // json['location'] = location
+                            // save the returned data of restaurants and events extracted in the previous .then
+                            // into the props which can be passed down to restaurant and event containers
+                            // to have their properties read into the cards.
                             return props.changeLocation(json)
                         });
                 }}
             >
-                {/* need input and button */}
+                {/* input field and button to be placed on nav-bar for searching various locations */}
 
                 <input type="text" placeholder="Enter Location!"></input>
                 <button type="submit">Submit Location</button>
