+++ package.json
@@ -29,7 +29,8 @@
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-router": "^6.10.0",
-		"react-router-dom": "^6.10.0"
+		"react-router-dom": "^6.10.0",
+		"sass-loader": "^13.2.2"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.21.4",
