+++ webpack.config.js
@@ -14,6 +14,9 @@ module.exports = {
 		}),
 	],
 	mode: process.env.NODE_ENV,
+	resolve: {
+		extensions: ['.js', '.jsx'],
+	},
 	module: {
 		rules: [
 			{
@@ -28,7 +31,7 @@ module.exports = {
 			},
 			{
 				test: /\.s?css/,
-				use: ['style-loader', 'css-loader'],
+				use: ['style-loader', 'css-loader', 'sass-loader'],
 			},
 		],
 	},
